Add tests from canonical mk implementation
I _did_ comment out some of the disequality tests since they seem to
redefine rembero and distincto. I'm not actually sure if I've
implemented distincto incorrectly here, but the rembero that this egg
ships is based on the rembero from the Reasoned Schemer, since that's
what most folk will expect.

Not sure what to do about that, open to someone submitting a patch
though!